NU-2091 handling advanced parameters errors
